Moved parent formstate handling up a level.
